fixed after code review
