+++ src/utils/compare-random.js
