+++ src/utils/get-random-array-element.js
@@ -1,0 +1,1 @@
+export const getRandomArrayElement = (array) => array[Math.floor(Math.random() * array.length)];
